Move symbol table building to symbol checker file
